+++ cases/example/login_test.js
@@ -1,10 +1,9 @@
 import { strict as assert } from 'assert';
 
 import gnhAccount from '../../steps/gnhAccount';
-import store from '../../steps/store';
 
 
-describe('test id', () => {
+describe('test auth', () => {
 	it('Зарегистрироваться на grassnearhome и перейти в профиль', () => {
 		gnhAccount.open('https://grassnearhome.ru/login');
 		const headerProfile = gnhAccount.auth();
@@ -13,17 +12,6 @@ describe('test id', () => {
 			headerProfile,
 			originHeader,
 			'Страница профиля успешно открыта',
-		)
-	});
-
-	it('Открыть первый магазин и добавить первый товар в корзину', () => {
-		store.open('https://grassnearhome.ru');
-		const foodSign = store.order();
-		const added = "Добавлено!";
-		assert.strictEqual(
-			foodSign,
-			added,
-			'Первый продукт успешно добавлен в корзину!',
 		)
 	});
 });
